xong xu li loi va update colum nest and table nest cho user

+++ app/Http/Controllers/UserController.php
@@ -6,12 +6,14 @@ use App\Http\Requests\StoreUserRequest;
 use App\Http\Requests\UpdateUserRequest;
 use Illuminate\Http\Request;
 use App\Models\Group;
+use App\Models\Nest;
 use App\Models\User;
 use App\Services\Interfaces\BorrowServiceInterface;
 use App\Services\Interfaces\DeviceServiceInterface;
 use Illuminate\Support\Facades\DB;
 use App\Services\Interfaces\UserServiceInterface;
 use App\Services\Interfaces\GroupServiceInterface;
+use App\Services\Interfaces\NestServiceInterface;
 use Illuminate\Support\Facades\Auth;
 
 
@@ -21,36 +23,41 @@ class UserController extends Controller
     protected $groupService;
     protected $borrowService;
     protected $deviceService;
+    protected $nestService;
 
-    public function __construct(UserServiceInterface $userService, GroupServiceInterface $groupService, BorrowServiceInterface $borrowService, DeviceServiceInterface $deviceService,)
+    public function __construct(UserServiceInterface $userService, NestServiceInterface $nestService, GroupServiceInterface $groupService, BorrowServiceInterface $borrowService, DeviceServiceInterface $deviceService,)
     {
         $this->groupService = $groupService;
         $this->userService = $userService;
         $this->borrowService = $borrowService;
         $this->deviceService = $deviceService;
+        $this->nestService = $nestService;
     }
     public function index(Request $request)
     {
-        if(!Auth::user()->hasPermission('User_viewAny')){
+        if (!Auth::user()->hasPermission('User_viewAny')) {
             abort(403);
         }
         $items = $this->userService->all($request);
         $groups = Group::get();
+        $nests = Nest::get();
         $param =
             [
                 'items' => $items,
                 'request' => $request,
                 'groups' => $groups,
+                'nests' => $nests,
             ];
         return view('users.index', $param);
     }
     public function create()
     {
-
         $groups = Group::get();
+        $nests = Nest::get();
         $params =
             [
                 'groups' => $groups,
+                'nests' => $nests,
             ];
         return view('users.create', $params);
     }
@@ -63,10 +70,12 @@ class UserController extends Controller
     public function edit($id)
     {
         $groups = Group::get();
+        $nests = Nest::get();
         $item = $this->userService->find($id);
         $params =
             [
                 'groups' => $groups,
+                'nests' => $nests,
                 'item' => $item,
             ];
         return view('users.edit', $params);
@@ -80,12 +89,17 @@ class UserController extends Controller
     public function destroy($id)
     {
         try {
+            if ($this->userService->isUserBorrow($id)) {
+                return redirect()->back()->with('error', 'Người dùng đang mượn thiết bị, không thể xóa!');
+            }
+
             $this->userService->destroy($id);
             return redirect()->route('users.index')->with('success', 'Xóa thành công!');
         } catch (\Exception $e) {
             return redirect()->back()->with('error', 'Xóa thất bại!');
         }
     }
+
     public function show($id)
     {
         $item = $this->userService->find($id);
@@ -94,7 +108,7 @@ class UserController extends Controller
     public function trash(Request $request)
     {
         $users = $this->userService->trash($request);
-        return view('users.trash', compact('users','request'));
+        return view('users.trash', compact('users', 'request'));
     }
     public function restore($id)
     {
@@ -152,6 +166,6 @@ class UserController extends Controller
             ->where('u.id', $id);
         $history = $queryBuilder->paginate(5);
         // dd($history);
-        return view('users.history', compact('user', 'history','changeStatus','changeApproved'));
+        return view('users.history', compact('user', 'history', 'changeStatus', 'changeApproved'));
     }
 }
